Change exported values to differentiate using from NPM and from CDN
Prior to this change, the default exported value, the global singleton, was exported along the error classes. This caused the global singleton plug to be available at `croct.default` at the browser instead of `croct`.

With this change, the javascript generated for the browser will export only the singleton plug as a global variable `croct` and the NPM package will export an object as module with the error classes and the singleton plug as its properties.